+++ src/main/java/com/github/gaboso/module/impl/NpmModule.java
@@ -1,8 +1,8 @@
 package com.github.gaboso.module.impl;
 
-import com.github.gaboso.format.Formatter;
-import com.github.gaboso.helper.CommandHelper;
+import com.github.gaboso.helper.CommandExecutor;
 import com.github.gaboso.module.Module;
+import com.github.gaboso.module.ModuleTypeEnum;
 
 import java.io.File;
 import java.util.Arrays;
@@ -14,20 +14,30 @@ import java.util.Objects;
  */
 public class NpmModule implements Module {
 
-    @Override
-    public boolean isProject(File[] listFiles) {
-        return Arrays.stream(listFiles)
-                     .filter(Objects::nonNull)
-                     .filter(file -> !file.isDirectory())
-                     .map(File::getName)
-                     .anyMatch("package.json"::equals);
+    private final CommandExecutor commandExecutor;
+
+    public NpmModule(CommandExecutor commandExecutor) {
+        this.commandExecutor = commandExecutor;
     }
 
     @Override
-    public void executeCommands(String projectName, String projectPath) {
-        Formatter formatter = new Formatter("NPM", projectName);
+    public boolean isProject(File[] listFiles) {
+        return Arrays.stream(listFiles)
+            .filter(Objects::nonNull)
+            .filter(file -> !file.isDirectory())
+            .map(File::getName)
+            .anyMatch("package.json"::equals);
+    }
 
-        CommandHelper.executeCMD(projectPath, "npm install && npm update", formatter);
+    @Override
+    public void executeCommands(String projectPath) {
+        String command = this.getType().getCommand();
+        commandExecutor.executeCMD(projectPath, command);
+    }
+
+    @Override
+    public ModuleTypeEnum getType() {
+        return ModuleTypeEnum.NPM;
     }
 
 }